% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{json_protocol}
\alias{json_protocol}
\title{JSON protocol implementation to connect dstack.ai API server.}
\usage{
json_protocol(server, error = .error)
}
\arguments{
\item{server}{A server to connect.}

\item{error}{An error handling function.}
}
\description{
JSON protocol implementation to connect dstack.ai API server.
}
