% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{push}
\alias{push}
\title{Pushes all commits to server. In the case of auto_push mode it sends only a total number
of elements in the frame. So call this method is obligatory to close frame anyway.}
\usage{
push(frame)
}
\arguments{
\item{frame}{A frame to push.}
}
\value{
Stack URL.
}
\description{
Pushes all commits to server. In the case of auto_push mode it sends only a total number
of elements in the frame. So call this method is obligatory to close frame anyway.
}
\examples{
\dontrun{
library(ggplot2)
library(dstack)
image <- qplot(clarity, data = diamonds, fill = cut, geom = "bar")
frame <- create_frame(stack = "diamonds")
frame <- commit(frame, image, "Diamonds bar chart")
print(push(frame)) # print actual stack URL
}
}
