% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{push_frame}
\alias{push_frame}
\title{Creates frame in the stack, commits and pushes data in a single operation.
        stack: A stack you want to commit and push to.}
\usage{
push_frame(
  stack,
  obj,
  description = NULL,
  params = NULL,
  profile = "default",
  handler = ggplot_handler(),
  protocol = NULL,
  config = yaml_config(),
  encryption = no_encryption
)
}
\arguments{
\item{stack}{A name of stack to use.}

\item{obj}{Object to commit and push, e.g. plot.}

\item{description}{Optional description of the object.}

\item{params}{Optional parameters.}

\item{profile}{Profile you want to use, i.e. username and token. Default profile is 'default'.}

\item{handler}{Specify handler to handle the object, if it's None then \code{ggplot_handler} will be used.}

\item{protocol}{Protocol to use, usually it is \code{NULL} it means that \code{json_protocol} will be used.}

\item{config}{Configuration to manage profiles. If it is unspecified \code{yaml_config} will be used.}

\item{encryption}{Encryption method by default \code{no_encryption} will be used.}
}
\value{
Stack URL.
}
\description{
Creates frame in the stack, commits and pushes data in a single operation.
        stack: A stack you want to commit and push to.
}
\examples{
\dontrun{
library(ggplot2)
library(dstack)
image <- qplot(clarity, data = diamonds, fill = cut, geom = "bar")
push_frame("diamonds", image, "Diamonds bar chart")
}
}
